animatediff-backend/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
--extra-index-url https://download.pytorch.org/whl/cu118
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.24.0
xformers>=0.0.22
opencv-python>=4.8.0
pillow>=10.0.0
pydantic>=2.4.0
python-multipart>=0.0.6
aiofiles>=23.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
redis>=5.0.0
celery>=5.3.0
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.0.0
safetensors>=0.4.0
huggingface-hub>=0.19.0
omegaconf>=2.3.0
imageio>=2.33.0
imageio-ffmpeg>=0.4.9